+++ frb_codegen/src/utils.rs
@@ -48,7 +48,7 @@ pub fn get_symbols_if_no_duplicates(configs: &[crate::Opts]) -> Result<Vec<Strin
     let mut explicit_raw_symbols = Vec::new();
     let mut all_symbols = Vec::new();
     for config in configs {
-        let raw_ir_file = config.get_ir_file();
+        let raw_ir_file = config.get_ir_file()?;
 
         // for checking explicit api duplication
         explicit_raw_symbols.extend(raw_ir_file.funcs.iter().map(|f| f.name.clone()));